Fix the installation of indata2json Python wrapper
For reasons that elude me, _something_ in the Python build system,
on _some_ systems (e.g. my Ubuntu 22.04 laptop, but not the Ubuntu 22.04
GitHub runners) refuses to install src/vmecpp/cpp/third_party as part of
the Python package.

After too much time trying to figure out what or why is responsible,
and after having verified empirically that it's really the fact that
the directory is called third_party that is to blame, I resorted to
simply move the Python bindings to our generic _util.py module.